Fixes for keychain
