+++ internal/cli/credstore/credstore.go
@@ -4,10 +4,10 @@ import "github.com/99designs/keyring"
 
 // Keyring provides the uniform interface over the underlying backends.
 type Keyring interface {
-	// Get returns an Item matching the key or ErrKeyNotFound
+	// Get returns an Item matching the key.
 	Get(key string) (keyring.Item, error)
-	// Set stores an Item on the keyring
+	// Set stores an Item on the keyring.
 	Set(item keyring.Item) error
-	// Remove removes the item with matching key
+	// Remove removes the item with matching key.
 	Remove(key string) error
 }
